requirements.txt
streamlit>=1.32.0
google-generativeai>=0.5.0
fpdf2>=2.7.9
